Converted all Matrix4ds from OpenGLDOTNet.Math to System.Numerics 4x4. Consequently converted all doubles to floats.

+++ RayTracer-App/aux_classes/Color.cs
@@ -7,19 +7,19 @@ desc: class that represents a RGB color as percents
 
 using System;
 using OpenGLDotNet;
-
+//CONVERTED DOUBLE -> FLOAT
 public class Color
 {
     //TODO DEFINE STATIC CONSTANTS FOR SPHERE COLOR, FLOOR COLOR, AND BACKGROUND COLOR
     const int COLOR_MAX = 255;
 
-    private double _r;
-    private double _g;
-    private double _b;
+    private float _r;
+    private float _g;
+    private float _b;
 
-    public double r { get => this._r; set => this._r = value; }
-    public double g { get => this._g; set => this._g = value; }
-    public double b { get => this._b; set => this._b = value; }
+    public float r { get => this._r; set => this._r = value; }
+    public float g { get => this._g; set => this._g = value; }
+    public float b { get => this._b; set => this._b = value; }
 
     //default constructor.. THE BACKGROUND COLOR
     public Color() 
@@ -29,7 +29,7 @@ public class Color
         this._b = 0;
     }
 
-    public Color( double r, double g, double b)
+    public Color( float r, float g, float b)
     {
         this._r = r;
         this._g = g;
